+++ cli/src/commands/mod.rs
@@ -1,7 +1,10 @@
+use std::env;
+
 /// Contains the types and functions to support the Enrich Command.
 pub mod enrich;
 /// Contains the types and functions to support the SBOM Command.
 pub mod sbom;
+mod snyk;
 
 
 // /// Extracts the value of an environment variable
