+++ core/src/clients/mod.rs
@@ -1,2 +1,4 @@
 mod snyk;
+pub mod github;
+
 pub use snyk::*;
