+++ core/src/services/mod.rs
@@ -1,7 +1,6 @@
 mod team;
 pub use team::*;
 
-mod pilot;
 pub mod providers;
 
 pub use pilot::*;
