+++ core/src/services/providers/github/mongo.rs
@@ -1,30 +1,59 @@
-use mongodb::{
-    Client as MongoClient,
-    Collection,
-    Database
-};
-use mongodb::bson::doc;
+use std::collections::HashMap;
+use std::fmt::{Debug, Formatter};
+use std::sync::Arc;
+
 use serde::{
     Deserialize,
     Serialize
 };
+use platform::Error;
 
 use platform::mongodb::{
+    Service as MongoService,
     Context as MongoContext,
+    MongoDocument,
+    Store,
+    client_from_context,
     mongo_doc,
-    MongoDocument
 };
 
-use crate::commands::github::GhProviderError;
 use crate::services::github::GhProviderError;
 use crate::services::providers::github::GhProviderError;
+
+pub struct GitHubProviderMongoService {
+    pub(crate) name: String
+}
+
+impl MongoService<GitHubSbomProviderEntry>
+    for GitHubProviderMongoService {
+    fn store(&self) -> Arc<Store> {
+        Arc::new(
+            Store::new(
+                &MongoContext {
+                    host: "localhost".to_string(),
+                    username: "root".to_string(),
+                    password: "harbor".to_string(),
+                    db_name: "harbor".to_string(),
+                    key_name: "last_hash".to_string(),
+                    port: 0,
+                }
+            )
+        )
+    }
+}
+
+impl Debug for GitHubProviderMongoService {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "GitHubProvider, name: {}", self.name)
+    }
+}
 
 /// Struct to define a GitHub Provider document in Mongo
 ///
 #[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct GitHubProviderDocument {
+pub struct GitHubSbomProviderEntry {
 
-    /// Unique id of the [GitHubProviderDocument].
+    /// Unique id of the [GitHubSbomProviderEntry].
     pub id: String,
 
     /// Url of the GitHub Repository
@@ -42,13 +71,13 @@ pub struct GitHubProviderDocument {
     /// Harbor v1 codebase id
     pub codebase_id: String,
 }
-mongo_doc!(GitHubProviderDocument);
+mongo_doc!(GitHubSbomProviderEntry);
 
 /// Function to update the last commit hash in a document
 ///
 pub async fn update_last_hash_in_mongo(
-    document: GitHubProviderDocument,
-    collection: Collection<GitHubProviderDocument>,
+    document: GitHubSbomProviderEntry,
+    collection: Collection<GitHubSbomProviderEntry>,
     last_hash: String,
 ) {
 
@@ -98,4 +127,45 @@ pub async fn get_mongo_db() -> Result<Database, GhProviderError> {
 
     // Get a handle to the database.
     Ok(client.database(&ctx.db_name))
+}
+
+#[tokio::test]
+async fn test_add_document() {
+
+    let svc = GitHubProviderMongoService {
+        name: String::from("test-name")
+    };
+
+    let id = String::from("test-id");
+    let repo_url = String::from("test-url");
+    let last_hash = String::from("test-last-hash");
+    let team_id = String::from("test-team-id");
+    let project_id = String::from("test-project-id");
+    let codebase_id = String::from("test-codebase-id");
+
+    svc.insert(
+        &mut GitHubSbomProviderEntry {
+            id,
+            repo_url,
+            last_hash,
+            team_id,
+            project_id,
+            codebase_id,
+        }
+    );
+
+    match svc.find(id.as_str()) {
+        Ok(opt) => match opt {
+            Some(doc) => {
+                assert_eq!(id, doc.id);
+                assert_eq!(repo_url, doc.repo_url);
+                assert_eq!(last_hash, doc.last_hash);
+                assert_eq!(team_id, doc.team_id);
+                assert_eq!(project_id, doc.project_id);
+                assert_eq!(codebase_id, doc.codebase_id);
+            },
+            None => panic!("No value in Option: Missing GitHubSbomProviderEntry")
+        },
+        Err(err) => panic!("Error getting GitHubSbomProviderEntry: {}", err)
+    }
 }
